Added tests for nextcloud controller
